spatiotemporal_data_cube_visualization/requirements.txt
pandas>=1.3.0
geopandas>=0.10.0
matplotlib>=3.4.0
numpy>=1.20.0
pyproj>=3.0.0
pyyaml>=6.0
shapely>=1.8.0